refactor(workflow-tasks): move component definition/recipe models into library
